+++ NEWS.md
@@ -1,3 +1,8 @@
+# MIRit 0.99.2
+
+MIRit now allows to filter the pathways used for topological analysis based
+on the number of nodes.
+
 # MIRit 0.99.1
 
 Functional enrichment analyses and TAIPA now use cached databases to reduce